findBreadcrumbEntries: Add handleMissing option
Add `handleMissing` option to `findBreadcrumbEntries`. When true, if
`activeKey` is not found in `graph`, the function returns an empty array
instead of throwing an exception.

A use case for this is a 404 error page with
`eleventyExcludeFromCollections` enabled, but using a base template that
includes `eleventyNavigationBreadcrumb`. By default, the error page
would fail to render because it would look for the 404 page in the
graph and not be able to find it.